Change successful "unlock" reply
Currently, the reply for successful "/v1/steady-state" requests is misleading:

> unlocked reboot lease for <OWNER_ID>

This response is sent even when the machine ID is NOT the holder of the lease
(which means that the lease is NOT released).

This commit changes the reply to use a generic message "steady-state confirmed"
(which is what Airlock logs in that case).