Change state of lock to unlocked even if holder not found
Currently, the `fleetlock_lock_state` transitions to 0 only if:
* the unlock handler is called and
* a lease with holder matching the id of the machine exists

However, if the lease itself is deleted, a call to the unlock
handler for the same group would succeed, but not change the
metric value for `fleetlock_lock_state` metric. This is not
the expected behavior.

This PR changes that behavior so that the metric does change
its value to 0, reflecting the reality that another machine
may obtain a lease on the group. i.e. it is "unlocked".